triple_buffer
Struct Output

pub struct Output<T: Send> { /* private fields */ }

Consumer interface to the triple buffer

The consumer of data can use this struct to access the latest published update from the producer whenever he likes. Readout is nonblocking: a collision between the producer and consumer will result in cache contention, but deadlocks and scheduling-induced slowdowns cannot happen.

Implementations

impl<T: Send> Output<T>

pub fn read(&mut self) -> &T

Access the latest value from the triple buffer

pub fn updated(&self) -> bool

Tell whether an updated value has been submitted by the producer

This method is mainly intended for diagnostics purposes. Please do not let it inform your decision of reading a value or not, as that would effectively be building a very poor spinlock-based double buffer implementation. If what you truly need is a double buffer, build yourself a proper blocking one instead of wasting CPU time.

pub fn peek_output_buffer(&self) -> &T

Query the current value of the output buffer

This is simply a read-only version of output_buffer_mut(). Please read the documentation of that method for more information on the precautions that need to be taken when accessing the output buffer in place.

In particular, remember that this method does not update the output buffer automatically. You need to call update() in order to fetch buffer updates from the producer.

pub fn output_buffer(&mut self) -> &mut T

๐Deprecated: Please use output_buffer_mut() instead

Access the input buffer directly

This is, for now, a deprecated alias to output_buffer_mut(). Please use this method instead.

In a future major release of this crate, output_buffer() will undergo a breaking change to instead provide read-only access, like peek_output_buffer() currently does. At that point, peek_output_buffer() will be deprecated.

Finally, in a later major release peek_output_buffer() will be removed.

The aim of this process is to eventually migrate towards the standard accessor()/accessor_mut() method naming convention that most Rust libraries follow.

pub fn output_buffer_mut(&mut self) -> &mut T

Access the output buffer directly

This advanced interface allows you to modify the contents of the output buffer, so that you can avoid copying the output value when this is an expensive process. One possible application, for example, is to post-process values from the producer before use.

However, by using it, you force yourself to take into account some implementation subtleties that you could normally ignore.

First, keep in mind that you can lose access to the current output buffer any time read() or update() is called, as it may be replaced by an updated buffer from the producer automatically.

Second, to reduce the potential for the aforementioned usage error, this method does not update the output buffer automatically. You need to call update() in order to fetch buffer updates from the producer.

pub fn update(&mut self) -> bool

Update the output buffer

Check if the producer submitted a new data version, and if one is available, update our output buffer to use it. Return a flag that tells you whether such an update was carried out.

Bear in mind that when this happens, you will lose any change that you performed to the output buffer via the output_buffer_mut() interface.